plugins/spell_check/requirements.txt
pyenchant
